support larger cardinality with .exhaustive()
